Continued Styling
Continued adding color and type treatment and began fiddling with
adequate spacing. Need to finalize the image hover with multiply
effect. Need to figure out issue with ‘slash.svg’, styling buttons in
the form, and image control in the project page. Spacing in the project
page (a-fan only) is still a work in progress.